Altered the Values class so that it does not inherit a List but exposes an IList interface. This approach enables the class to be constructed with an array rather than an enumerator reducing memory operations and duplication.
The Profile and Signature classes have been altered to create an array of values rather than using an enumerator via their accessor. This has the benefit of enabling the array to be allocated before values are added.

The Profile GetPropertyValues method has been altered to ensure that the start and end indexes are associated to the requested property before the enumeration starts. Array.BinarySearch is now used in place of a bespoke code when looking up value indexes for Profiles and Properties.

A static empty array is used when returning empty values for properties that don't have any values.

Signature now models Profile for caching and returning property values.

New tests have been added to iterate Profile and Signature property values.


Former-commit-id: eec5de5dc8565282b7655cbd8cf04091a197e75d