BUG/MAJOR: Random behaviour during automatic updates resolved
When the BinaryFile is moved ready to be replaced by the temporary data file it is now explicitly refreshed before any further checks or properties are retrieved. This refresh may not have completed in the prior version and as such random failure of an automatic update may have been experienced.