OPTIM/MINOR: Reduce memory consumption.
The Values property of the Signature class now returns an enumeration of available values rather than an array of values. This significantly reduces the amount of memory consumed compared to the original array implementation. The biggest benefit is gained when initialising the data set with the memory factory. The memory tests have been adjusted accordingly.

Introduced an integer list type to be used for handling integer values which is no longer derived from BaseEntity. This new list avoids the need to create class instances for each integer value retrieved.

The revised integer list also modifies GetRange so that the underlying array can be used without copying the elements that are required into a new array. The resulting list is returned as an IList interface. This change reduces memory consumption when operating in memory mode.

Removed display of ranked signature index cache status for tests as this data is no longer cachable.

Updated version number for reflect changes to integer lists and memory improvements.

Increased the tolerances associated with some of the performance tests for bad User-Agents.


Former-commit-id: e7d75df904891b208dd53f2923631ecaa506eb8d