EXAMPLE/BUILD: Added all profiles example, and removed unnecessary configurations.
The all profiles example fetches all the hardware profiles, then outputs all the devices and their properties to a CSV file. Two filters are applied to the properties that are printed: firstly only hardware properties are fetched as they are the only properties that relate to a hardware profiles, also any javascript property is given the value "N/A" as these can return long multiline javascript which will ruin any CSV file formatting.

In the same way as with the C repository, all configurations other than x64 have been removed as they are not needed and just make maintaing and adding projects more complex.


Former-commit-id: b24b87c17af537a8c157b165e7fdfc3f913e41d6