FEAT: WebProvider.Download now returns status code.
When the WebProvider is being used outside a web environment to make use of the update feature, it is useful to see the return code of the data file download. The codes are detailed in the LicenceKeyResults enumeration.
Updating manually in this way may mean that the Download method is called when there is no bin directory (where the API searches for a licence key file). For this reason a check has been added to the Keys method in LicenceKeys which checks for the existance of the bin directory before searching it for a licence key file.